+++ package.json
@@ -11,13 +11,11 @@
   "dependencies": {
     "@radix-ui/react-slot": "^1.1.0",
     "@tanstack/react-query": "^5.50.1",
-    "autoprefixer": "^10.4.19",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "lucide-react": "^0.407.0",
-    "next": "^14.2.5",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "14.2.4",
+    "react": "^18",
+    "react-dom": "^18",
     "sweetalert2": "^11.12.2",
     "tailwind-merge": "^2.4.0",
     "tailwindcss-animate": "^1.0.7",
@@ -32,17 +30,12 @@
     "@types/react-slick": "^0.23.13",
     "eslint": "^8",
     "eslint-config-next": "14.2.4",
-    "postcss": "^8.4.39",
+    "postcss": "^8",
     "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
-<<<<<<< HEAD
-    "tailwindcss": "^3.4.4",
-    "typescript": "5.5.3"
-=======
     "react-slick": "^0.30.2",
     "slick-carousel": "^1.8.1",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
->>>>>>> 1a4ccecbc8929c170a6fcd014341b685ffb5b9a0
+    "typescript": "5.5.3"
   }
 }
